[fix] 총대폼 수정 route
