+++ src/components/Comment.js
@@ -1,14 +1,15 @@
 import { useEffect, useState } from "react";
 import { dbService, firebaseInstance } from "../fbase";
 import Commentlist from "./Commentlist";
-
+import { useParams } from "react-router-dom";
 
 const Comment= ({ userObj,qnaObj,detailObj,isOpener }) => {
+  const {id} = useParams();
     const [comment, setComment] = useState("");
     const [comments, setComments] = useState([]);
     const timestamp=firebaseInstance.firestore.FieldValue.serverTimestamp;
     useEffect(() => {
-      dbService.doc(`startlist/${detailObj.id}`).collection("QnA").doc(`${qnaObj.id}`).collection("comments").orderBy('createdAt').onSnapshot((snapshot) => {
+      dbService.doc(`startlist/${id}`).collection("QnA").doc(`${qnaObj.id}`).collection("comments").orderBy('createdAt').onSnapshot((snapshot) => {
           const commentArray = snapshot.docs.map((doc) => (
             {
             id: doc.id,
@@ -20,7 +21,7 @@ const Comment= ({ userObj,qnaObj,detailObj,isOpener }) => {
   
       const onSubmitComment = async (event) => {
           event.preventDefault();
-          await dbService.doc(`startlist/${detailObj.id}`).collection("QnA").doc(`${qnaObj.id}`).collection("comments").add({
+          await dbService.doc(`startlist/${id}`).collection("QnA").doc(`${qnaObj.id}`).collection("comments").add({
             text: comment,
             createdAt: timestamp(),
             creatorId: userObj.uid,
